% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvgrid.R
\name{csvgrid-shiny}
\alias{csvgrid-shiny}
\alias{csvgridOutput}
\alias{renderCsvgrid}
\title{Shiny bindings for csvgrid}
\usage{
csvgridOutput(outputId, width = "100\%", height = "500px")

renderCsvgrid(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Must be valid CSS units (e.g. \code{"100\%"},
\code{"400px"}) or numbers (interpreted as pixels).}

\item{expr}{An expression that generates a \code{\link{csvgrid}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})?
Useful when wrapping this function inside another.}
}
\description{
Output and render functions for using \code{\link{csvgrid}} inside a Shiny
app or interactive R Markdown document.
}
